Net: refactor HostAddress class to its own set of files.
